Lucas/Proposta/requirements.txt
selenium==4.16.0
pandas>=2.1.3
openpyxl==3.1.2
webdriver-manager==4.0.1
requests==2.31.0
urllib3==2.1.0
numpy==2.1.3
python-docx==1.1.2